requirements.txt
numpy
wandb
tqdm
Pillow
scikit-learn
matplotlib
pandas
torchvision==0.15.2
torch==2.0.1